IOS-2132 Migrate all block related methods to List service
