+++ Anytype/Sources/PresentationLayer/Modules/HomeWidgets/Widgets/Tree/Specific/ObjectTreeWidgetModuleAssembly.swift
@@ -27,7 +27,7 @@ final class ObjectTreeWidgetModuleAssembly: HomeWidgetCommonAssemblyProtocol {
             subscriptionManager: serviceLocator.treeSubscriptionManager(),
             pageRepository: serviceLocator.pageRepository(),
             documentsProvider: serviceLocator.documentsProvider,
-            blockActionsService: serviceLocator.blockActionsServiceSingle(),
+            blockActionsService: serviceLocator.blockListService(),
             output: output
         )
