+++ Anytype/Sources/PresentationLayer/TextEditor/EditorPage/EditorPageBlocksStateManager.swift
@@ -76,7 +76,7 @@ final class EditorPageBlocksStateManager: EditorPageBlocksStateManagerProtocol {
 
     private let document: BaseDocumentProtocol
     private let modelsHolder: EditorMainItemModelsHolder
-    private let blockActionsServiceSingle: BlockActionsServiceSingleProtocol
+    private let blockActionsService: BlockListServiceProtocol
     private let toastPresenter: ToastPresenterProtocol
     private let actionHandler: BlockActionHandlerProtocol
     private let pasteboardService: PasteboardServiceProtocol
@@ -94,7 +94,7 @@ final class EditorPageBlocksStateManager: EditorPageBlocksStateManagerProtocol {
         document: BaseDocumentProtocol,
         modelsHolder: EditorMainItemModelsHolder,
         blocksSelectionOverlayViewModel: BlocksSelectionOverlayViewModel,
-        blockActionsServiceSingle: BlockActionsServiceSingleProtocol,
+        blockActionsService: BlockListServiceProtocol,
         toastPresenter: ToastPresenterProtocol,
         actionHandler: BlockActionHandlerProtocol,
         pasteboardService: PasteboardServiceProtocol,
@@ -107,7 +107,7 @@ final class EditorPageBlocksStateManager: EditorPageBlocksStateManagerProtocol {
         self.document = document
         self.modelsHolder = modelsHolder
         self.blocksSelectionOverlayViewModel = blocksSelectionOverlayViewModel
-        self.blockActionsServiceSingle = blockActionsServiceSingle
+        self.blockActionsService = blockActionsService
         self.toastPresenter = toastPresenter
         self.actionHandler = actionHandler
         self.pasteboardService = pasteboardService
@@ -333,7 +333,7 @@ final class EditorPageBlocksStateManager: EditorPageBlocksStateManagerProtocol {
         AnytypeAnalytics.instance().logReorderBlock(count: movingBlocksIds.count)
         
         Task { @MainActor in
-            try await blockActionsServiceSingle.move(
+            try await blockActionsService.move(
                 contextId: document.objectId,
                 blockIds: movingBlocksIds,
                 targetContextID: targetId,
